stocks/requirements.txt
django
djangorestframework
qrcode